Add missing columns to the tasks table and edit columns functionality
The tasks table now includes all columns, with the ones not displaying by default hidden. You can change the columns shown using the Edit columns menu, except for the Task and Recipe columns, which constitute the core of the table and must always be shown. That's why they are checked and disabled in the Edit columns menu.

Signed-off-by: Belen <belenbarrospena@gmail.com>