Apply correct class to error headings and add help icon
Apply the .error class to the error headings in the "Recent builds" section so that they have the right hover state. Also, add the help icon to all table headings and remove references to project-build-time.html, which will be renamed to build-time.html to keep consistency with the information architecture documentation.

Signed-off-by: Belen <belenbarrospena@gmail.com>