Configuration page and linking to external sources
This commit mainly adds the new configuration page (build-configuration.html) and provides a mechanism to add links to external websites in order to extend the information provided about a certain element. Examples are linking layers to the information about them in the OpenEmbedded layer index, and linking BitBake variable descriptions to the full explanation provided in the Yocto Project reference manual.